Get rid of borrows from Objects/genobject.c
Summary:
Add a PyAPI declaration to the functions we need and remove the borrows from
them and their dependencies.

This is slightly less trivial than previous borrow cleanup diffs because we
still have one function whose signature we override via a wrapper, and
therefore need to keep the cpp directives to support the wrapper code.

Reviewed By: itamaro

Differential Revision: D65371388

fbshipit-source-id: 29f2eb3643dc7b04709be292ed2fc9e58b7d23d8