Make CinderX work with optimized Python builds (again)
Summary:
D62604340 introduced a step to the build where `sed` is used to edit `site.py` to *enable* CinderX in the 3.12 distribution as needed. Unfortunately this doesn't work in optimized builds. Those builds "deep-freeze" a number of startup Python libraries including `site.py` in a state *before* the edit is made.

To fix this, move the CinderX initialization code into a Python module that's not in the frozen libraries, and only include this in the distribution if CinderX is desired. During Python startup, if the `init_cinderx` does not exist then proceed without it leaving the set of imported modules unchanged. This resolves the issue D62604340 was trying to address.

Reviewed By: itamaro

Differential Revision: D65584596

fbshipit-source-id: 212dbe22a93fa44d33750166b80d03bffa559502