Patch test_posixpath for huge `pwd.getpwall()` output
Summary: In certain environments, `pwd.getpwall()` can return very long lists. Having the test iterate on the full list isn't very useful for the purpose of the test (it's just testing the same thing over and over), but can dramatically increase the runtime of the test (as long as ~10m).

Reviewed By: parulgupta1004

Differential Revision: D69377194

fbshipit-source-id: 36dbd37eb31f228f514cf6e46ee91b8efe4a1519