Avoid strong references that need to be released at shutdown for code data
Summary:
Currently we have a `CodeData` object which has strong references to some data about where a code object came from when we have a nested code object.

We can end up releasing these at shutdown in 3.12 after the runtime has exited and we don't hold the GIL which results in a fatal error (P1743316681).

Instead of holding a strong reference to the globals/builtins/module we can instead hold a borrowed ref to the function (which holds a strong reference to the fields of interest). We can then clear our references when the function is destroyed.

To support this we need to clear the code's `co_code` after we've delivered the event, which matches the behavior in 3.12: https://www.internalfb.com/code/fbsource/[e39e710ea330c14dbb75e86d8bcf1ba094df00f0]/third-party/python/3.12/Objects/funcobject.c?lines=857

Reviewed By: alexmalyshev

Differential Revision: D70411784

fbshipit-source-id: 7beeb3b199e355b46fadcfc4a6b799ddb2a8c531