Don't replace the existing object when module is already imported
Summary:
2nd attempt of D70011859

This one differs in that we delay the decision on not replacing. There's some additional logic for choosing to do an eager import if the parent needs to be loaded eagerly (https://www.internalfb.com/code/fbsource/[655f0007c024f0c0dbbfa2724b56881bf9e7ef24]/third-party/python/3.10.cinder/Python/import.c?lines=1701). It's odd that this breaks things as the module should already be loaded, but this is causing some subtle incompatibility.

If we've eagerly imported a module but later the loader says we should lazily import it we can potentially replace the object that is on the module as part of updating the lazy module state: https://www.internalfb.com/code/fbsource/[03b897a2919ec8626bdd1f61293ff6cb1bd75dba]/third-party/python/3.10.cinder/Python/import.c?lines=1751.

This is fine if we're just replacing the module with the module, but if the module has already been replaced with something else then we'll discard the old thing.

If we change this to saying the module is already available and doesn't need to be lazily imported we will trigger more imports than normal - that's because when we process the from list children could be imported that would have succeeded as lazy imports.

So we should just avoid doing updates to the lazy module state once the value is published in `sys.modules`.

Reviewed By: brittanyrey

Differential Revision: D70667959

fbshipit-source-id: b7188a019caa969e8a1a700d61cddeaf9826894d