Avoid strong references that need to be released at shutdown for code data
Summary:
Fixed from previous version: We need to clear the code objects `jit_reg_units` in addition to `jit_code_outer_funcs` in `funcDestroyed`

Currently we have a `CodeData` object which has strong references to some data about where a code object came from when we have a nested code object.

We can end up releasing these at shutdown in 3.12 after the runtime has exited and we don't hold the GIL which results in a fatal error (P1743316681).

Instead of holding a strong reference to the globals/builtins/module we can instead hold a borrowed ref to the function (which holds a strong reference to the fields of interest). We can then clear our references when the function is destroyed.

Reviewed By: alexmalyshev

Differential Revision: D70674730

fbshipit-source-id: 24472dec5f9837ebd1e518803579c05fedaf7d4d