Don't recurse up to 10 times
Summary:
This check doesn't seem to be required to be 10. When we hit recursion on an imported module it should immediately be a cycle, it seems that typically this just causes us to recurse deeper until we finally report the error anyway.

Further simplification could be done here to just have `_lazy_state_add` be a check to see if the value is present.

Reviewed By: alexmalyshev

Differential Revision: D68841472

fbshipit-source-id: 2348acccb282036e9a0fc41fdd61c1e314fe96e1