gh-97514: Don't use Linux abstract sockets for multiprocessing (GH-98501) (GH-98503)
Summary:
Backporting this fix to multiprocessing/connection.py from Python 3.10.9.

**Below is copied from GitHub**

Linux abstract sockets are insecure as they lack any form of filesystem
permissions so their use allows anyone on the system to inject code into
the process.

This removes the default preference for abstract sockets in
multiprocessing introduced in Python 3.9+ via
#18866 while fixing
#84031.

Explicit use of an abstract socket by a user now generates a
RuntimeWarning.  If we choose to keep this warning, it should be
backported to the 3.7 and 3.8 branches.
(cherry picked from commit 49f6106)

Automerge-Triggered-By: GH:gpshead

**End summary copied from GitHub**

Reviewed By: jermenkoo, ip4368

Differential Revision: D71409753

fbshipit-source-id: aa75b3f95f6f0469f613127568c6967b337aed52

Co-authored-by: Gregory P. Smith <greg@krypto.org>