Always include init_cinderx (take #3)
Summary:
**Attempt #3 for D71057259 and D71057259 (folded into this).**

First attempt (D70213609) was reverted because it triggered stack overflows when
the Cinder interpreter loop loaded import cycles w/ Lazy Imports enabled. This is
fixed in D68841472.

Second attempt (D71057259) was reverted because it led to a large number of ASAN
errors and timeouts when running ASAN checks.  This was traced back to
D70661206, which has been reverted in D71503509.

Reviewed By: itamaro

Differential Revision: D71586058

fbshipit-source-id: fc5802e8ded59c293e59f1c42e8193ff0ef11b16