Replace asyncio.tasks.get_async_stack with asyncio.ci_get_async_stack
Summary:
Nothing we have actually uses `get_async_stack` and it's really just left-over from a PR we tried and failed to get upstreamed. In this diff we replace it with `ci_get_async_stack` which both makes it clearer it's not a standard Python feature and functions slightly differently.

This is all implemented in 3.12 directly rather than in CinderX as the ability to reconstruct async stacks is actually not tied to CinderX in any way. It also will hopefully not be needed in future Python versions as a new upstream version of this feature should be coming.

The functional change is we now return the full stack - all the way to the Python entry-point, rather than stopping at the bottom of the task stack. This helps with implementing a backward compatibility library for stack walking functions we had in 3.10.cinder but which don't exist in 3.12 CinderX (see following changes). The implementation is in C as this is actually a lot easier due to functions to find the generator object, if any, associated with a frame.

Reviewed By: martindemello

Differential Revision: D73524378

fbshipit-source-id: e52298f808b6740339de4b566a272e93d9d05fc0