gh-132969 fix ProcessPoolExecutor exception and hang
Summary:
Github issue that I reported https://github.com/python/cpython/issues/132969 contains a repro of an exception and a hang in Python's `concurrent.futures.ProcessPoolExecutor` when `shutdown(wait=False)` is used in conjunction with a couple of other pre-conditions.

This diff
(1) adds python unittest of that repro, and
(2) fixes the ProcessPoolExecutor to fix the bug.

Reviewed By: blchxfm

Differential Revision: D73898605

fbshipit-source-id: 61eb472df46bbddfafd59e56aa0f5009789efefe