Fix sending to AwaitableValue
Summary:
In Cinder Python 3.6/8 we added a fast way to "send" (in the generator sense) values to objects. This was upstreamed and made generally available in 3.10. However, when porting `_asynciomodule` it looks like some of the old infra for our 3.8 implementation was left behind and `AwaitableValue` continued to use this. The old implementation isn't actually used though so sending to `AwaitableValue` always took the slower path.

Sadly this didn't actually improve things in a Swayze experiment but we might as well keep the cleaned up code.

Reviewed By: martindemello

Differential Revision: D76005584

fbshipit-source-id: a5dad39f51a5d12684b4d76c4cd6ce6d1b5aa79f