Backport gh-116402: Avoid readline in test_builtin TTY input tests
Summary:
upstream issue: https://github.com/python/cpython/issues/116402
upstream PR: https://github.com/python/cpython/pull/122447

`test_builtin.test_input_tty` is consistently hanging on macOS when running the full test suite in a single worker, due to an issue with `readline` being imported earlier in the same process.

this upstream patch fixes that by temporarily "detaching" readline. here's the upstream PR description:

The intention here is to make the TTY `input()` tests run in as many scenarios as possible. For this to happen, we need to temporarily remove the function pointer set to `call_readline` in `PyInit_readline()`. This pointer is normally `NULL` without `readline` and `PyOS_Readline` uses a basic implementation of its own in this case. This is what we want.

We could try and avoid importing `readline` in tests, but we have played this whack-a-mole for a long time. For instance, now that `pdb` uses `rlcompleter`, `test_builtin` imports that always (since it imports `doctest` in `load_tests()` and `doctest` imports `pdb` as it subclasses it). I think we should stop trying to not load `readline`, and just detach its function temporarily. This avoids discussions of "unimporting" readline, which would be problematic as `PyInit_readline()` sets global state on the C side through `setup_readline`.

While CI is famously not run with a TTY and those tests will be skipped there, people running tests sequentially from their terminal will now be able to have those tests run in many more situations than before.

I did the function pointer unsetting with `ctypes` for simplicity. If that's unavailable, the context manager simply skips the tests.

Reviewed By: bowiechen

Differential Revision: D77306493

fbshipit-source-id: fd9816286865ffd9cdbc3ceb7371402df23a2725