Reduce initialization required of _PyInterpreterFrame in JIT
Summary:
This is the first step in reducing the cost of our JIT interpreter frames on 3.12. This eliminates the initialization of many of the fields inside of JIT interpreter frames. Instead we set `f_funcobj` to a helper object which will perform the initialization when called and this will finish up the initialization. The runtime will call into it using the hooks in D76850761.

This is only enabled with the `ENABLE_LIGHTWEIGHT_FRAMES` feature so that we can run on non-meta Python builds that don't have the feature.

Reviewed By: alexmalyshev

Differential Revision: D76775590

fbshipit-source-id: 2b94194d64f934ca77fcec7cf811c398254be7bb