Improve opcode metadata generation
Summary:
For the adaptive static Python interpreter support I modified our opcode generation (D75254549) so that we modified CPython's script and output different names. That's not going to work so well on 3.14 where the opcode generation is entirely done from the interpreter loop and involves a more complicated library instead of a single script.

This moves to using macros to define the names from the originally generated file instead, which should work better across versions.

Reviewed By: alexmalyshev

Differential Revision: D77602847

fbshipit-source-id: d88218f6687ef02555a8cf0275f4db8445cdcd09