Resource Tracker Cache Cleanup Check
Summary:
More context can be found [here](https://fb.workplace.com/groups/1069402721675548/permalink/1132909071991579/) and in the previous [diff](https://www.internalfb.com/diff/D79196602).

The general idea is that multiprocess library fails with bundled runtimes, which we have a fix for in the previous diff. However, because the resource tracker is spawned behind the scenes this fix does not stop it from crashing. Because it crashes it does not write the proper data to the cache and so when it tries to clean up keys it throws an [error](https://www.internalfb.com/phabricator/paste/view/P1886824837). The simple fix for this is to just simply check if that key exists for a given type in the cache. With this fix you get a successful/crashless [run](https://www.internalfb.com/phabricator/paste/view/P1886924225)

Reviewed By: alexmalyshev

Differential Revision: D79268625

fbshipit-source-id: fad33cb57fb5a7846f46b574da8e682b63c83004