Reifier returns function
Summary:
This updates the protocol between the runtime and CinderX for not fully reified frames. Previously the contract was that we'd update `f_funcobj` when we initialized the frame. This has a drawback in that we don't get any future calls to update the frame while it's still under our control.

So this instead returns the current function from the vectorcall implementation so that the runtime can have access to it when it needs it. When we finish with execution of a function if the frame is reified we'll publish the real function in at as we're no longer involved.

Reviewed By: alexmalyshev

Differential Revision: D78124674

fbshipit-source-id: ad6bbe0caf5302037e9fe59565af408a261b9276