Make 3.12 invalidate cached __lazy_imports_enabled__ on updates to lazy import settings
Summary:
This is potentially one of the key differences between the 3.12 and 3.10.cinder implementations. In 3.10.cinder there's a counter which is maintained that causes all of the cached lazy import state to be updated between modifications to the lazy import settings. In 3.12 once we've cached it we're done. This could potentially cause issues if something like:

main.py:
```
import foo
import bar
```
foo:
```
add_main_to_excluding(...)
```

Here on 3.10 `bar` would presumably be eagerly imported because disabling lazy imports would get disabled and the cache sequence would be invalidated. On 3.12 `bar` would still proceed lazily because we'd only consult `__lazy_imports_enabled__` which has already been set.

I have no idea how critical this is / likely it is to happen, but the sequence was seemingly added in 3.10 for some reason. It's hard to track this down because the commit in Cinder 3.10 was a backport from "main" and the commit in Cinder 3.8 was a backport from Cinder 3.10.

Reviewed By: itamaro

Differential Revision: D80116250

fbshipit-source-id: bc61ed12f143551e45f20bdb8194836a2c7c76d0