Add patch 'add-parallel-gc-hooks'
Summary:
This diff adds a patch to the `3.14` Meta-internal fork.

Adds the parallel GC hooks - these enable us to plug in the CinderX parallel GC. There was discussion of upstreaming the parallel GC at the core dev sprints so we may be able to stop maintaining this patch in the future but there's not yet a concrete plan.

This is a port of D75002643. It's mostly identical except some small changes on where it actually plugs into the collect function. It is not currently plugged into the free-threaded GC.

Reviewed By: itamaro

Differential Revision: D84398790

fbshipit-source-id: 8fabb7b107e6413a88871214e380a6d206a92863