Add patch 'tls-model-initial-exec'
Summary:
Use the `internal-exec` thread local storage model when building Meta's
internal fork of CPython as a shared library. This makes Python about 5.5%
faster on the pyperformance test suite. The known risk is that if lots of
libraries took this approach, the application could run out of available
internal-exec TLS slots and fail to load.

Reviewed By: itamaro

Differential Revision: D84916347

fbshipit-source-id: 5c735502b781efc8e3db658d76a9574c46091eea