Fix build issues caused by Lazy Imports 3.14t
Summary:
The land of Lazy Imports in 3.14t (D85977460) led to Github test failures (e.g. T244893771).  [This](https://github.com/facebookincubator/cinder/actions/runs/19307995749) is the link to the builds/tests run on the change.
Although some of the issues were pre-existing, some of them were due to the Lazy Imports
3.14t change:
- Windows build errors ([example](https://github.com/facebookincubator/cinder/actions/runs/19307995749/job/55220483261)): These seem likely due to not defining `_Py_atomic_load_uchar_acquire()` and `_Py_atomic_store_uchar_release()` in `pyatomic_msc.h`
- macOS (free-threading) build error ([link](https://github.com/facebookincubator/cinder/actions/runs/19307995749/job/55220483210)): This is likely due to having a label immediately followed by a declaration in `dictobject.c`, which is not allowed sometimes in C (I think before C-23).

Reviewed By: itamaro

Differential Revision: D86888783

fbshipit-source-id: 0ccd7e86e17ac7a05ddc6f39b96022741d9b6312